Updated ScheduleControllerTests
